Move the php-selector plugin to a must-use
