+++ provisioning/roles/nginx/templates/etc/nginx/conf.d/error-page
@@ -4,6 +4,6 @@ error_page 500 @error_page;
 location @error_page {
     # If encountered an error and using the backend cookie, let's show that special error page
     if ( $cookie_backend ) {
-        rewrite ^       /wp-content/plugins/php-selector/error.html      break;
+        rewrite ^       /wp-content/mu-plugins/php-selector/error.html      break;
     }
 }
