+++ provisioning/roles/wordpress/tasks/main.yml
@@ -123,5 +123,16 @@
   with_items:
     - "{{ domain }}"
 
-- name: "Copy php-selector plugin into place"
-  copy: src=php-selector dest="{{ wp_doc_root }}/{{ enviro }}/wp-content/plugins/" owner={{ web_user }} group={{ web_user }} mode=0644
+- name: Must use plugin directory
+  file: path="{{ wp_doc_root }}/{{ enviro }}/wp-content/mu-plugins" state=directory owner={{ web_user }} group={{ web_group }}
+
+- name: Copy php-selector plugin into place
+  copy: src=php-selector dest="{{ wp_doc_root }}/{{ enviro }}/wp-content/mu-plugins/" owner={{ web_user }} group={{ web_user }} mode=0644
+
+- name: Enable php-selector must use plugin
+  copy:
+    src: "{{ wp_doc_root }}/{{ enviro }}/wp-content/mu-plugins/php-selector/php-selector.php"
+    dest: "{{ wp_doc_root }}/{{ enviro }}/wp-content/mu-plugins/"
+    owner: "{{ web_user }}"
+    group: "{{ web_user }}"
+    mode: 0644
